Move service naming handling business logic to the Integration class
