+++ bridge/_files_integrations.php
@@ -6,7 +6,6 @@ return [
     __DIR__ . '/../src/Integrations/Integrations/AbstractIntegrationConfiguration.php',
     __DIR__ . '/../src/Integrations/Integrations/DatabaseIntegrationHelper.php',
     __DIR__ . '/../src/Integrations/Integrations/HttpClientIntegrationHelper.php',
-    __DIR__ . '/../src/Integrations/Integrations/SpanTaxonomy.php',
     __DIR__ . '/../src/Integrations/Integrations/DefaultIntegrationConfiguration.php',
     __DIR__ . '/../src/Integrations/Integrations/Integration.php',
     __DIR__ . '/../src/Integrations/Integrations/AMQP/AMQPIntegration.php',
