+++ src/Integrations/Integrations/ElasticSearch/V1/ElasticSearchIntegration.php
@@ -3,7 +3,6 @@
 namespace DDTrace\Integrations\ElasticSearch\V1;
 
 use DDTrace\Integrations\Integration;
-use DDTrace\Integrations\SpanTaxonomy;
 use DDTrace\SpanData;
 use DDTrace\Tag;
 use DDTrace\Type;
@@ -42,7 +41,7 @@ class ElasticSearchIntegration extends Integration
                 }
 
                 $span->name = "Elasticsearch.Client.__construct";
-                SpanTaxonomy::handleInternalSpanServiceName($span, ElasticSearchIntegration::NAME);
+                Integration::handleInternalSpanServiceName($span, ElasticSearchIntegration::NAME);
                 $span->type = Type::ELASTICSEARCH;
                 $span->resource = "__construct";
                 $span->meta[Tag::COMPONENT] = ElasticSearchIntegration::NAME;
@@ -105,7 +104,7 @@ class ElasticSearchIntegration extends Integration
         \DDTrace\trace_method('Elasticsearch\Endpoints\AbstractEndpoint', 'performRequest', function ($span) {
             $span->name = "Elasticsearch.Endpoint.performRequest";
             $span->resource = 'performRequest';
-            SpanTaxonomy::handleInternalSpanServiceName($span, ElasticSearchIntegration::NAME);
+            Integrations::handleInternalSpanServiceName($span, ElasticSearchIntegration::NAME);
             $span->type = Type::ELASTICSEARCH;
             $span->meta[Tag::SPAN_KIND] = 'client';
             $span->meta[Tag::COMPONENT] = ElasticSearchIntegration::NAME;
@@ -125,7 +124,7 @@ class ElasticSearchIntegration extends Integration
         \DDTrace\trace_method('Elasticsearch\Connections\Connection', 'performRequest', function ($span, $args) {
             $span->name = "Elasticsearch.Endpoint.performRequest";
             $span->resource = 'performRequest';
-            SpanTaxonomy::handleInternalSpanServiceName($span, ElasticSearchIntegration::NAME);
+            Integrations::handleInternalSpanServiceName($span, ElasticSearchIntegration::NAME);
             $span->type = Type::ELASTICSEARCH;
             $span->meta[Tag::SPAN_KIND] = 'client';
             $span->meta[Tag::COMPONENT] = ElasticSearchIntegration::NAME;
@@ -170,7 +169,7 @@ class ElasticSearchIntegration extends Integration
                     }
 
                     $span->meta[Tag::SPAN_KIND] = 'client';
-                    SpanTaxonomy::handleInternalSpanServiceName($span, ElasticSearchIntegration::NAME);
+                    Integrations::handleInternalSpanServiceName($span, ElasticSearchIntegration::NAME);
                     $span->type = Type::ELASTICSEARCH;
                     $span->resource = ElasticSearchCommon::buildResourceName($name, isset($args[0]) ? $args[0] : []);
                     $span->meta[Tag::COMPONENT] = ElasticSearchIntegration::NAME;
@@ -189,7 +188,7 @@ class ElasticSearchIntegration extends Integration
             $operationName = str_replace('\\', '.', "$class.$name");
             $span->name = $operationName;
             $span->resource = $operationName;
-            SpanTaxonomy::handleInternalSpanServiceName($span, ElasticSearchIntegration::NAME);
+            Integrations::handleInternalSpanServiceName($span, ElasticSearchIntegration::NAME);
             $span->type = Type::ELASTICSEARCH;
             $span->meta[Tag::COMPONENT] = ElasticSearchIntegration::NAME;
         });
@@ -211,7 +210,7 @@ class ElasticSearchIntegration extends Integration
 
             $span->name = "Elasticsearch.$namespace.$name";
             $span->resource = ElasticSearchCommon::buildResourceName($name, $params);
-            SpanTaxonomy::handleInternalSpanServiceName($span, ElasticSearchIntegration::NAME);
+            Integrations::handleInternalSpanServiceName($span, ElasticSearchIntegration::NAME);
             $span->type = Type::ELASTICSEARCH;
             $span->meta[Tag::COMPONENT] = ElasticSearchIntegration::NAME;
         });
