+++ src/Integrations/Integrations/Guzzle/GuzzleIntegration.php
@@ -5,7 +5,6 @@ namespace DDTrace\Integrations\Guzzle;
 use DDTrace\Http\Urls;
 use DDTrace\Integrations\HttpClientIntegrationHelper;
 use DDTrace\Integrations\Integration;
-use DDTrace\Integrations\SpanTaxonomy;
 use DDTrace\SpanData;
 use DDTrace\Tag;
 use DDTrace\Type;
@@ -38,7 +37,7 @@ class GuzzleIntegration extends Integration
             function (SpanData $span, $args, $retval) use ($integration) {
                 $span->resource = 'send';
                 $span->name = 'GuzzleHttp\Client.send';
-                SpanTaxonomy::handleInternalSpanServiceName($span, GuzzleIntegration::NAME);
+                Integration::handleInternalSpanServiceName($span, GuzzleIntegration::NAME);
                 $span->type = Type::HTTP_CLIENT;
                 $span->meta[Tag::SPAN_KIND] = Tag::SPAN_KIND_VALUE_CLIENT;
                 $span->meta[Tag::COMPONENT] = GuzzleIntegration::NAME;
@@ -81,7 +80,7 @@ class GuzzleIntegration extends Integration
             function (SpanData $span, $args, $retval) use ($integration) {
                 $span->resource = 'transfer';
                 $span->name = 'GuzzleHttp\Client.transfer';
-                SpanTaxonomy::handleInternalSpanServiceName($span, GuzzleIntegration::NAME);
+                Integration::handleInternalSpanServiceName($span, GuzzleIntegration::NAME);
                 $span->type = Type::HTTP_CLIENT;
                 $span->meta[Tag::SPAN_KIND] = Tag::SPAN_KIND_VALUE_CLIENT;
                 $span->meta[Tag::COMPONENT] = GuzzleIntegration::NAME;
