+++ src/Integrations/Integrations/Memcache/MemcacheIntegration.php
@@ -4,7 +4,6 @@ namespace DDTrace\Integrations\Memcache;
 
 use DDTrace\Integrations\DatabaseIntegrationHelper;
 use DDTrace\Integrations\Integration;
-use DDTrace\Integrations\SpanTaxonomy;
 use DDTrace\Obfuscation;
 use DDTrace\SpanData;
 use DDTrace\Tag;
@@ -145,7 +144,7 @@ class MemcacheIntegration extends Integration
     {
         $span->name = "Memcache.$command";
         $span->type = Type::MEMCACHED;
-        SpanTaxonomy::handleInternalSpanServiceName($span, MemcacheIntegration::NAME);
+        Integration::handleInternalSpanServiceName($span, MemcacheIntegration::NAME);
         $span->resource = $command;
         $span->meta['memcache.command'] = $command;
         $span->meta[Tag::SPAN_KIND] = 'client';
