+++ src/Integrations/Integrations/SQLSRV/SQLSRVIntegration.php
@@ -5,7 +5,6 @@ namespace DDTrace\Integrations\SQLSRV;
 use DDTrace\HookData;
 use DDTrace\Integrations\DatabaseIntegrationHelper;
 use DDTrace\Integrations\Integration;
-use DDTrace\Integrations\SpanTaxonomy;
 use DDTrace\SpanData;
 use DDTrace\Tag;
 use DDTrace\Type;
@@ -207,7 +206,7 @@ class SQLSRVIntegration extends Integration
         $span->name = $name;
         $span->resource = $query ?? $name;
         $span->type = Type::SQL;
-        SpanTaxonomy::handleInternalSpanServiceName($span, SQLSRVIntegration::NAME);
+        Integration::handleInternalSpanServiceName($span, SQLSRVIntegration::NAME);
         $span->meta[Tag::SPAN_KIND] = 'client';
         $span->meta[Tag::COMPONENT] = SQLSRVIntegration::NAME;
         $span->meta[Tag::DB_SYSTEM] = SQLSRVIntegration::SYSTEM;
